(maint) Fix symlinks on windows
This commit removes the attempt to use the win32-dir gem to make
junctions in Ruby on windows in favor of using mklink directly. This gem
is in the process of being deprecated across Puppet for many reasons,
the relevant reason here being that it uses methods that are not
supported in constrained OS like Windows Server Core or
Windows Containers. Similar projects like [rspec-puppet](https://github.com/rodjek/rspec-puppet/blob/master/lib/rspec-puppet/setup.rb#L163)
have followed this approach as well.